* The last changes didn't seem to work for windows. I don't think I can do this:
#if defined(UINT32)

because that's a typedef, not a header definition. I don't know if that's
really the problem them.

I changed the definition to test for WIN32 and not MinGW at the same
time, which I've been trying to avoid. Let's see if that works.

Curiously, CPAN Testers shows that 0.06_02 has all passes, so what
the heck am I changing all this for?


git-svn-id: http://brian-d-foy.svn.sourceforge.net/svnroot/brian-d-foy/trunk@1943 d9f60a02-fc28-0410-9752-bc0d6c4f57a6